+++ lua/user/configs/editor/snacks.lua
@@ -19,5 +19,13 @@ return function()
 		bufdelete = { enabled = true },
 		lazygit = { enabled = true },
 		image = { enabled = true },
+		input = { enabled = true },
+		picker = { enabled = true },
+		terminal = { enabled = true },
+		words = { enabled = true },
+		quickfile = { enabled = true },
+		scope = { enabled = true },
+		scroll = { enabled = true },
+		statuscolumn = { enabled = true },
 	})
 end